Refactor of retriever for multiple tape-aware dsmc
Previously, control flow looped with blocking calls to dsmc.
Now we fork off up to a limit (maxretrievers) calls to dsmc and poll
them for completion on each iteration of the main loop.

We also prevent jobs from being issued that uses a tape that's recorded
as being in use by a job.

The code is ready for accepting tape names from request files, but no
such requests exist yet.