config: Rename config items only affecting a single component.
There is a need to extend the configuration with multiple space
threshold levels for tsmarchiver, and while breaking things we might as
well do it properly.

Variables only affecting a single component are prepended with
archiver_ or retriever_ as appropriate.

Warning messages are emitted when old configuration directives are seen,
and silently converted to new directives to avoid breaking things too much.