Add desc-short and desc-long to config
Allows to configure descriptions of this ENDIT instance.

In combination with centralized logging and statistics gathering
this can be used to keep track of which instance did what.

An envisioned use is to set desc-short to be identical as the dCache
hsmInstance, and use desc-long to emit instance-specific metadata that
can be picked up by a central analysis script.