tsm_getvolumecontent: Don't skip damaged files
Skipping damaged files is counterproductive. tsmretriever assigns files it
can't map to the tape default which doesn't abide remount limits etc due to
some tricky corner cases. This leads to tsmretriever doing a rather tight
loop when trying to retrieve damaged files.

Mapping damaged files properly allows the remount limit to kick in, and also
gives an idea of which tape is broken in the ENDIT logs instead of having
to dive into logs on the TSM server.