Add tsmtapehints.pl, create tape hint file on pool-node
Script to create tape hint file, to be run on pool node in similar
manner to the ENDIT daemons, using the same configuration file.

dsmc query archive -detail provides enough information to create a tape
hint file. In fact, this output also has info on relative position on
tape which prompts us to store the file in JSON format instead of
an ad-hoc text format.

Compared to tsm_getvolumecontent.pl this method adds the benefit of being
able to run on a pool-node without additional TSM configuration and also
provides details on restore order which can prove useful if the need arises
to do chunked retrieves to easily allow for higher-priority requests to
skip ahead of the queue without having to wait for too long.

Another big benefit is the ability to do additional sanity checks,
currently only detection of dCache files being stored multiple times in TSM
has been added.

A downside compared to tsm_getvolumecontent.pl is that this method only
provides a numeric volume ID that is hard to translate into a physical
volume, but that information is likely only of interest for the TSM
administrator when debugging eventual issues.

Another more obvious issue is this method being slower, approximately
doubling the hint file creation time.

Fixes #14.