tsm_getvolumecontent: Don't skip damaged files(2)
The fix in commit f866fb00869ace235ae61260c9165ef5c9d2d55d to not skip
damaged files turned out to be incomplete.

If a volume was flagged access=unavailable or similar then files on it
would not be included and we would see very frequent retries.

This commit changes the volume list retrieval to only ignore volumes that
are empty or are pending (ie. empty/reclaimed but on deletion hold due to
storage pool reusedelay!=0).

Should also reduce the impact of #21