tsmdeleter: Implement queuing and periodic queue processing
This revamps tsmdeleter to add incoming deletion requests to a queue and
periodically process that queue.

While not a perfect solution to the problem, it alleviates the impact
of deletions slowly trickling in over a longer period of time, as is often
the case with WLCG deletion campaigns.

Adds config item deleter_queueprocinterval that understands simple keywords
like hourly/daily/monthly, but also crontab style time definitions if
the Perl module Schedule::Cron is installed.

NOTE: The deletion queue adds stateful long-lived data on the pool,
which is also reflected in the README.

Solves #22