Retry instead of exiting if writing file list fails
Running out of space can happen on pools where the dcache pool size is
set too high, not leaving enough margins for ongoing restores.

Since we can easily retry, let's do that instead of exiting if we fail
to write a file list. If the space shortage is temporary, this allows
the system to recover without intervention.