tsmarchiver: Actually delete listfile when job done
It was thought that the UNLINK flag to tempfile() would delete the file
when it went out of scope, but that only applies to filehandle-only
mode.  When returning a filename, the file is deleted when the program
exits.

To avoid filling up the requestlists dir while running, ensure we delete
the listfile when a worker is finished.