retriever: Gracefully handle incomplete/broken request files
Apparently the way the dCache ENDIT plugin writes request files can
cause the tsmretriever to barf on an invalid JSON file, for example if
the request file is opened for writing but content isn't written yet.

Work around this by catching errors and retrying a few times with 100ms
delay inbetween.