retriever: Apply brake when fs fills up
Version 2 of the dCache ENDIT plugin has evolved from preallocating
space for files to allocating space as files are processed. This moves
the responsibility to the tsmretriever to avoid the filesystem becoming
full.

This adds tunables and checks to avoid filling the fs in two steps:
1) Avoid starting new workers when we notice a backlog.
2) If we grow short of space, kill all running workers.

Three config knobs are introduced to allow tuning the behavior:

retriever_buffersize:
The buffer (in GiB) on filesystem reserved for the ENDIT retriever,
defaults to 1000 GiB.

retriever_backlogthreshold:
Threshold (in %), when filesystem use of the ENDIT reserved buffersize
goes above this don't start new workers. Defaults to 10% (of the
buffersize, NOT the total fs usage).

retriever_killthreshold:
Threshold (in %), when filesystem use of the ENDIT reserved buffersize
goes above this running workers are killed to avoid filling the
filesystem. Defaults to 95% (of the buffersize, NOT the total fs
usage).