Add cputime ulimit to spawned dsmc processes
dsmc has a few corner case bugs when it can get stuck in a loop
consuming 100% cpu and never recover.

Work around this by adding a cputime limit to spawned dsmc:s, 48
cpuhours should be enough to not trigger in normal usecases.

Since there seems to be no perl module for this packaged in
distributions use the kludge of running ulimit in the dsmc invocation.

Solves: https://github.com/neicnordic/endit/issues/27