retriever: Cache request info instead of parsing on each iteration
With the late allocation (version 2) dCache ENDIT plugin there is a
higher probability of us seeing lots of files in the request directory.

Scalability tests shows that with O(100k) files the time to parse the
JSON request files on each iteration becomes noticeable.

So instead we cache the parsed data, and ensure that we invalidate the
cached data if the files change.

Also, it's more efficient to iterate over (large) hashes with a
while ... each construct instead of foreach ... keys, the first uses an
iterator while the second actually constructs a temporary list.