archiver: Use lower timeout when retrying
Introduce archiver_retrytimeout (default 3600 s) that is used instead of
archiver_timeout in retry situations.

This improves the behaviour if store fails for some reason (for example
due to TSM server downtime/outage), and a long archiver_timeout causes
the store to time out on a dCache level before a retry happens.

Closes: https://github.com/neicnordic/endit/issues/24