Prometheus stats: Emit HELP strings for all metrics
Issues can arise when having multiple ENDIT instances on the same host
due to the node exporter generating conflicting HELP strings if they are
missing.

Work around this by emitting HELP strings ourselves. Note that they
aren't meant to be helpful for humans, and thus they're rather terse.

A non-functional change is changing the the logsuffix internal
variable to logname allowing it to be used in more places.