Lower default retriever_buffersize to 100 GiB
As of dCache ENDIT provider 2.0.0-alpha.10 the staging process has been
rearranged to do allocation as soon as a file has begun staging instead
of after staging was finished. This reduces the space needed to cater
for doing late allocation.

To be on the safe side, reduce the default from 1000 GiB to 100 GiB to
leave plenty of room for stray files etc.